Originally, this github project ran on localhost. This commit includes changes to the web.xml, persistence.xml and "test" folder that allows for this project to run over the internet. I omitted some details for security.
Jelastic was used to host this application server in the cloud on a Glassfish 5.1 server with a Postgres 12.0 database. This application could also be deployed in a docker container onto AWS or Azure if you can find a good Glassfish 5.1 image and configure it correctly to run. I found Jelastic to be easier, as it allowed for easier container configuration, although I am new to docker, so a more experienced user may find it easier to work with docker directly.